migrate single sim ci test  to launch_testing
