+++ controls/sae_2025_ws/src/uav/package.xml
@@ -18,6 +18,7 @@
 
   <test_depend>ament_copyright</test_depend>
   <test_depend>python3-pytest</test_depend>
+  <test_depend>launch_testing</test_depend>
   <depend>uav_interfaces</depend>
   <export>
     <build_type>ament_python</build_type>
